requirements.txt
numpy==1.14.1
pandas==0.21.0
sqlalchemy==1.2.4
psycopg2-2.7.4-cp35-cp35m-win_amd64.whl
flask==0.12.2
Shapely-1.6.4.post1-cp35-cp35m-win_amd64.whl
requests==2.18.4
pytest==3.4.2